contingency tables use doubles so weighting doesn't need to round
commit 9c71f906eaa2e5a18179e036fc162e14535c3266
Author: Ryan Suderman <ryants@ku.edu>
Date:   Thu Dec 31 12:19:35 2015 -0600

    ct with doubles reweight unit test

commit a955e8245b8f3d0bb2746441e6cfadf0c2bff48b
Author: Ryan Suderman <ryants@ku.edu>
Date:   Wed Dec 30 22:22:31 2015 -0600

    modified contingency tables to use Doubles

+++ src/main/scala/infcalcs/Other.scala
@@ -20,10 +20,10 @@ object MathFuncs {
    * Converts frequencies to probabilities by normalizing each count by the
    * vector sum.
    */
-  def freqToProb: TraversableOnce[Int] => TraversableOnce[Double] =
+  def freqToProb: TraversableOnce[Double] => TraversableOnce[Double] =
     l => {
-      val s = l.sum.toDouble
-      if (s != 0) l map (x => (x / s).toDouble) else l map (x => 0.0)
+      val s = l.sum
+      if (s != 0) l map (x => x / s) else l map (x => 0.0)
     }
 
   def avg(t: Seq[Double]): Double = t.sum / t.length.toDouble
