+++ src/main/scala/infcalcs/tables/ImportedTable.scala
@@ -4,8 +4,8 @@ import infcalcs.IOFile._
 
 /** Class for reading contingency tables from a file. */
 class ImportedTable(fileName: String) extends ContTable {
-  val table: Vector[Vector[Int]] =
-    importData(fileName) map (x => x map (y => y.toInt))
+  val table: Vector[Vector[Double]] =
+    importData(fileName) map (x => x map (y => y.toDouble))
 
   lazy val rows: Int = table.length
   lazy val cols: Int = if (table.isEmpty) 0 else table(0).length
